Property Info Extraction Bot.
This Bot Operates on inserting a Links in a text File.
Program picks the link and Perform Operation and Extract Following Data:
1) Address
2) Price
3) Area
4) Price Per Meter sq.
5) Rooms
6) Property Dealer's Address
7) Property Dealer's Tel/ Mob Number
And Put All this Data into 2 Files. One with Rooms Greater then 6 and Other with Less 6. 
This Code is Highly Optimized and Extract 2 Property in 1 Second.
Exception Handlings, Errors has been watched carefully.